dev Question -> Parcelable; load result ConvertResultFunction -> i Function; extras DiagnosticTest  -> iSerializable; push to Start, Diagnostic, Done activity by  extras#getSerializable(s)

+++ app/src/main/java/ru/hardwork/onlinesocialdiagnosticapp/model/user/UserResult.java
@@ -40,7 +40,7 @@ public class UserResult {
         this.user = user;
     }
 
-    public long getDiagnosticId() {
+    public int getDiagnosticId() {
         return diagnosticId;
     }
